Task

+++ frontend/src/component/Filter/InquiryTableFiltered.js
@@ -94,8 +94,10 @@ const InquiryTableFlitered = ({ modalHandler, refresh }) => {
     }
 
     const delteHandler = async (id) => {
-        const data = await axios.delete(`/api/v1/inquiry/delete/${id}`);
-        fetchInquiry();
+        if(window.confirm("Are you sure ?")){
+            const data = await axios.delete(`/api/v1/inquiry/delete/${id}`);
+            fetchInquiry();
+        }
     }
