+++ frontend/src/component/Tables/Inquiry/inquiryTable.js
@@ -1,67 +1,78 @@
-import React, { useState, useEffect } from 'react';
-import { Link } from 'react-router-dom';
+import React, { useState, useEffect,useMemo } from 'react';
 import Styles from './inquiryTable.module.css'
 import Add from '../../../Assets/Add.svg'
-import MaterialTable from 'material-table';
 import FilterAltIcon from '@mui/icons-material/FilterAlt';
 import Modal from '../../Layout/Modal/Modal';
-import { Paper } from '@material-ui/core';
 import axios from 'axios';
-import CustomerEditForm from '../../Forms/CustomerEditForm';
-import DummyEditForm from '../../Forms/DummyEditForm';
-import { toast, ToastContainer } from 'react-toastify'
+import { toast } from 'react-toastify'
 import Select from 'react-select'
 import TextField from '@mui/material/TextField';
 import InquiryEditForm from '../../Forms/InquiryEditForm';
-const InquiryTable = ({ modalHandler ,modalHandler2,refresh}) => {
+import { dateformater,dateParser } from '../Utils/util';
+import FileDownloadIcon from '@mui/icons-material/FileDownload';
+import MaterialReactTable from 'material-react-table';
+import { ExportToCsv } from 'export-to-csv';
+
+import {
+  Box,
+  Button,
+  IconButton,
+  Tooltip,
+} from '@mui/material';
+import { Delete, Edit } from '@mui/icons-material';
+const InquiryTable = ({ modalHandler ,modalHandler2,refresh,isOpen}) => {
 
   const [inquiries, setInquiries] = useState([]);
-  const [tabledata, setTableData] = useState([])
+  const [originalData, setOriginalData] = useState([]);
+  const [tabledata, setTableData] = useState([]);
   const [editModal, setEditModal] = useState(false);
   const [editModalData, setEditModalData] = useState({});
   const [startDate, setStartDate] = useState(new Date('2022-08-01'));
   const [endDate, setEndDate] = useState(new Date());
   const [branches, setBranches] = useState([]);
-  let selectedBranch = [];
   const [isLoading, setIsLoading] = useState(false);
   const [salesman, setSalesman] = useState([]);
-  let selectedSalesman = [];
+  let [selectedBranch,setSelectedBranch] = useState(null);
+  let [selectedSalesman,setSelectedSalesman] = useState(null);
+
   const fetchSalesmen = async () => {
     const { data } = await axios.get("/api/v1/salesman/getall");
-
-    const salesmen = data.salesmans.map((branch) => (
+    const salesmen = data.salesmans.map((salesman) => (
       {
-        name: branch.name,
-        value: branch.name,
-        label: branch.name
+        name: salesman.name,
+        value: salesman.name,
+        label: salesman.name
 
       }
     ))
     setSalesman(salesmen);
   }
-  const fetchArchitectsofSalesman = async () => {
-    setIsLoading(true);
-    sleep(500);
 
+  const fetchBranches = async () => {
+    const { data } = await axios.get("/api/v1/branch/getall");
     
-    const response = await axios.post("/api/v1/salesman/inquiry", selectedSalesman, { headers: { "Content-Type": "application/json" } });
+    const branches = data.branches.map((branch) => (
+      {
+        branchname: branch.branchname,
+        value: branch.branchname,
+        label: branch.branchname
 
-    
-    const newarchitects = response.data.inquiries;
-
-    setTableData(newarchitects);
-    setIsLoading(false);
+      }
+    ))
+    setBranches(branches);
   }
-  const handlesalesman = (selected) => {
-    console.log(selected);
 
-    selectedSalesman = selected;
-    fetchArchitectsofSalesman();
+  const handlesalesman = (selected) => {
+    setSelectedSalesman(selected.value);
+    fetchFilteredInquiries(selected.value,selectedBranch);
+  }
+
+  const handlebranch = (selected) => {
+    setSelectedBranch(selected.value);
+    fetchFilteredInquiries(selectedSalesman,selected.value);
   }
 
   const modifyData = (data) => {
-
-
     let datass1 = data.map((d) => {
       if (d.architectTag) {
         return {
@@ -89,17 +100,18 @@ const InquiryTable = ({ modalHandler ,modalHandler2,refresh}) => {
       }
       return d
     })
-    console.log(datass1)
     return datass1
-
   }
 
   const delteHandler = async (id) => {
-    const data = await axios.delete(`/api/v1/inquiry/delete/${id}`);
-    fetchInquiry();
+    if( window.confirm(id)){      
+       const data1 = await axios.delete(`/api/v1/inquiry/delete/${id}`);
+       // fetchInquiry();
+       const { data } = await axios.get("/api/v1/inquiry/getall");
+       setOriginalData(data.inquiries);
+    }
+
   }
-
-
 
   const startDateHandler = (e) => {
     setStartDate(new Date(e.target.value));
@@ -109,85 +121,143 @@ const InquiryTable = ({ modalHandler ,modalHandler2,refresh}) => {
     setEndDate(new Date(e.target.value));
   }
 
-  const dateformater = (date) => {
-    let year = date.getFullYear();
-    let month = (date.getMonth() + 1) > 9 ? date.getMonth() + 1 : '0' + date.getMonth();
-    let day = (date.getDay() + 1) > 9 ? date.getDay() + 1 : '0' + date.getDay();
-    return `${year}-${month}-${day}`
+  const getInquiryData = (mobileno) => {
+    let inquiry = originalData.filter((item) => item.mobileno === mobileno);
+    setEditModalData(inquiry[0]);
+    setEditModal(true);
   }
 
-  const submitDateRangeHandler = (e) => {
-    console.log(startDate, endDate);
-    let data = inquiries.filter((item) => {
-      let date = item.date;
-      date = new Date(date);
-      if (date < endDate && date > startDate) {
-        return true
+  const submitDateRangeHandler = () => {
+
+    if(startDate && endDate){
+
+      let data = originalData.filter((item) => {
+        let date = (item.followupdate);
+        date = new Date(date);
+        
+        if(date){
+        if (date < endDate && date > startDate) {
+          return true
+        }
+        else {
+          return false
+        }
       }
-      else {
+      else{
         return false
       }
+      })
+      // setInquiries()
+      setTableData(getInquiry(data))
+
+    }
+
+  }
+
+
+  function getInquiry(data){
+    let inquires = data.map((item)=>{
+      return {
+        date:item.date,
+        name:item.name,
+        followupdate:item.followupdate,
+        stage:item.stage,
+        mobileno:item.mobileno,
+        requirement:item.requirement.map((req)=>req.requirement).join('-'),
+        salesmen:item.salesmen.map((req)=>req.name).join('-')
+      }
     })
-    console.log(data)
-    setTableData(data)
-    
+
+    return inquires;
+
+
   }
   
   const fetchInquiry = async () => {
     const { data } = await axios.get("/api/v1/inquiry/getall");
-    console.log(data)
-    setInquiries(modifyData(data.inquiries));
-    setTableData(modifyData(data.inquiries));
-  }
-
-  const fetchBranches = async () => {
-    const { data } = await axios.get("/api/v1/branch/getall");
-    
-    const branches = data.branches.map((branch) => (
-      {
-        branchname: branch.branchname,
-        value: branch.branchname,
-        label: branch.branchname
-
+    setOriginalData(data.inquiries);
+    let inquires = data.inquiries.map((item)=>{
+      return {
+        date:item.date,
+        name:item.name,
+        followupdate:item.followupdate,
+        stage:item.stage,
+        requirement:item.requirement.map((req)=>req.requirement).join('-'),
+        salesmen:item.salesmen.map((req)=>req.name).join('-'),
+        mobileno:item.mobileno
       }
-    ))
-    setBranches(branches);
+    })
+    console.log(modifyData(inquires))
+    setInquiries(modifyData(inquires));
+    setTableData(modifyData(inquires));
   }
+
+
   const sleep = time => {
     return new Promise(resolve => setTimeout(resolve, time));
   };
 
-  const fetchInquiriesofBranch = async () => {
-    setIsLoading(true);
-    sleep(500);
-    
-    console.log(selectedBranch);
-    const response = await axios.post("/api/v1/branch/inquiry", selectedBranch, { headers: { "Content-Type": "application/json" } });
-    
-    console.log(response);
-    const newcust = response.data.inquiries;
-    
-    setTableData(newcust);
-    setIsLoading(false);
-  }
-  const handlebranch = (selected) => {
-    console.log(selected);
-    
-    selectedBranch = selected;
-    fetchInquiriesofBranch();
-  }
 
+  const fetchFilteredInquiries =async(salesman, branch) => {
+
+    let filteredData = originalData.filter((item)=>{
+      let isBranch = false;
+      let isSalesman = false;
+
+      if(item.branches.length === 0 && branch===null){
+        isBranch = true;
+      }
+
+      if(item.salesmen.length === 0 && salesman===null){
+        isSalesman = true;
+      }
+      
+      item.branches.forEach((branchObject)=>{
+        if(Object.values(branchObject).includes(branch) || branch===null){
+        isBranch = true;
+      }})
+      item.salesmen.forEach((salesmanObj)=>{
+        if(Object.values(salesmanObj).includes(salesman) || salesman===null){
+          isSalesman = true;
+        }})
+
+      if(isSalesman && isBranch){
+        return true
+      }
+    })
+    let data = filteredData.map((item)=>{
+        return {
+          date:item.date,
+          name:item.name,
+          followupdate:item.followupdate,
+          stage:item.stage,
+          requirement:item.requirement.map((req)=>req.requirement).join('-'),
+          salesmen:item.salesmen.map((req)=>req.name).join('-'),
+          mobileno:item.mobileno
+        }
+      })
+    setInquiries(modifyData(data));
+    setTableData(modifyData(data));  
+  }
 
   useEffect(() => {
     fetchInquiry();
+    fetchFilteredInquiries(selectedSalesman, selectedBranch)
     fetchBranches();
     fetchSalesmen();
     
   }, [refresh]);
-  const handleCallbackCreate = (childData) => {
-    
+
+  useEffect(()=>{
+    fetchFilteredInquiries(selectedSalesman, selectedBranch);
+    // submitDateRangeHandler();
+  },[originalData])
+
+  const handleCallbackCreate = async(childData) => {
     toast.success("Inquiry edited");
-    fetchInquiry();
+    const { data } = await axios.get("/api/v1/inquiry/getall");
+    setOriginalData(data.inquiries);
+
   }
 
   const customStyles = {
@@ -217,7 +287,45 @@ const InquiryTable = ({ modalHandler ,modalHandler2,refresh}) => {
         padding: 0
     })
 };
+const columns = useMemo(
+  () => [
+    { header: 'Date', accessorKey: 'date', type: "date", dateSetting: { locale: "en-GB" }, 
+    Cell: ({cell})=>(dateformater(cell.getValue()))},
+    { header: 'Name', accessorKey: 'name' },
+    { header: 'Follow Update', accessorKey: 'followupdate', type: "date", dateSetting: { locale: "en-GB" }, Cell: ({cell})=>(dateformater(cell.getValue())) },
+    {header: 'Stage', accessorKey:'stage'},
+    {header: 'Requirement', accessorKey: 'requirement'},
+    {header: 'Salesman', accessorKey:'salesmen'},
+    { header: 'Mobile Number', accessorKey: 'mobileno' },
+  ],
+  [],
+);
+const ops = [
+  { header: 'Date', accessorKey: 'date', type: "date", dateSetting: { locale: "en-GB" }, Cell: ({cell})=>(dateformater(cell.getValue())) },
+  { header: 'Name', accessorKey: 'name' },
+  { header: 'Follow Update', accessorKey: 'followupdate', type: "date", dateSetting: { locale: "en-GB" },Cell: ({cell})=>(dateformater(cell.getValue()))  },
+  {header: 'Stage', accessorKey:'stage'},
+  {header: 'Requirement', accessorKey: 'requirement'},
+  {header: 'Salesman', accessorKey:'salesmen'},
+  { header: 'Mobile Number', accessorKey: 'mobileno' },
+]
+const csvOptions = {
+  fieldSeparator: ',',
+  quoteStrings: '"',
+  decimalSeparator: '.',
+  showLabels: true,
+  useBom: true,
+  useKeysAsHeaders: false,
+  headers: ops.map((c) => c.header),
+};
+const csvExporter = new ExportToCsv(csvOptions);
+const handleExportData = () => {
 
+  csvExporter.generateCsv(tabledata);
+};
+const handleExportRows = (rows) => {
+  csvExporter.generateCsv(rows.map((row) => row.original));
+};
   return (
     <div className={Styles.container}>
       <div className={Styles.table}>
@@ -270,78 +378,95 @@ const InquiryTable = ({ modalHandler ,modalHandler2,refresh}) => {
             <button className={Styles.SubmitButton} onClick={(e) => submitDateRangeHandler(e)} type="submit"> Submit </button>
           </div>
         </div>
+        {inquiries &&
+          <MaterialReactTable
+            displayColumnDefOptions={{
+              'mrt-row-actions': {
+                muiTableHeadCellProps: {
+                  align: 'center',
+                },
 
-        {inquiries && <MaterialTable
-          isLoading={isLoading}
-          className={Styles.Table}
-          columns={[
-            { title: 'Date', field: 'date', type: "date", dateSetting: { locale: "en-GB" }, },
-            { title: 'Name', field: 'name' },
-            { title: 'Email', field: 'Email', hidden: 'true' },
-            { title: 'Address', field: 'address' },
-            { title: 'Birth Date', field: 'birthdate', hidden: 'true' },
-            { title: 'Marriage Date', field: 'marriagedate', hidden: 'true' },
-            { title: 'Remarks', field: 'remarks', hidden: 'true' },
-            { title: 'Order Value', field: 'orderValue', hidden: 'true' },
-            { title: 'Sales Person', field: 'salesPerson', hidden: 'true' },
-            { title: 'Tag', field: 'tag' },
+                size: 120,
+              },
+            }}
 
-
-          ]}
-          data={tabledata}
-          options={{
-            sorting: true,
-            headerStyle: {
-              zIndex: 0
-            },
-            showTitle: false,
-            actionsColumnIndex: -1,
-            filtering: true,
-            exportButton:true
-
-          }}
-          components={{
-            Container: props => <Paper {...props}
-              elevation={0}
-              style={{
-                padding: 20,
-                width: "100%",
-              }} />
-          }}
-
-
-          actions={[
-            {
-              icon: 'edit',
-              tooltip: 'Edit',
-              onClick: (event, rowData) => {
-                window.scrollTo({
-                  top: 0,
-                  left: 0,
-                  behavior: "smooth"
-                });
-                setEditModalData(rowData);
-                setEditModal(!editModal);
-                console.log(`Edit `, rowData)
-              }
-            },
-            {
-              icon: 'delete',
-              tooltip: 'Delete',
-              onClick: (event, rowData) => {
-                window.scrollTo({
-                  top: 0,
-                  left: 0,
-                  behavior: "smooth"
-                });
-                
-                delteHandler(rowData._id);
-                console.log(`delete `, rowData)
-              }
+            muiTopToolbarProps={
+              ({ }) => ({
+                color: 'green',
+                sx: { display: 'block' },
+                zIndex: '0'
+              })
             }
-          ]}
+            columns={columns}
+            data={tabledata}
+            enableEditing
+            enableRowNumbers
+            rowNumberMode='original'
+            enableTopToolbar={!editModal && !isOpen}
 
-        />}
+            muiTablePaginationProps={{
+              rowsPerPageOptions: [200,300,400],
+              showFirstLastPageButtons: true,
+            }}
+            enableGlobalFilter={true}
+            positionActionsColumn='last'
+            renderRowActions={({ row, table }) => (
+              <Box sx={{ display: 'flex', gap: '1rem' }}>
+                <Tooltip arrow placement="left" title="Edit">
+                  <IconButton onClick={() => {
+                    window.scrollTo({
+                      top: 0,
+                      left: 0,
+                      behavior: "smooth"
+                    });
+                    getInquiryData(row.original.mobileno)
+                    // setEditModalData(row.original)
+                    // setEditModal(true);
+                  }}>
+                    <Edit />
+
+                  </IconButton>
+                </Tooltip>
+                <Tooltip arrow placement="right" title="Delete">
+                  <IconButton color="error" onClick={() => {
+                    window.scrollTo({
+                      top: 0,
+                      left: 0,
+                      behavior: "smooth"
+                    });
+                    delteHandler(row.original.mobileno);
+                  }}>
+                    <Delete />
+                  </IconButton>
+                </Tooltip>
+              </Box>
+            )}
+            renderTopToolbarCustomActions={({ table }) => (
+              <Box
+                sx={{ display: 'flex', gap: '1rem', p: '0.5rem', flexWrap: 'wrap' }}
+              >
+                <Button
+                  disabled={table.getPrePaginationRowModel().rows.length === 0}
+
+                  onClick={() =>
+                    handleExportRows(table.getPrePaginationRowModel().rows)
+                  }
+                  startIcon={<FileDownloadIcon />}
+                  variant="contained"
+                >Export All Rows</Button>
+                <Button
+                  className={Styles.bu}
+                  color="primary"
+                  onClick={handleExportData}
+                  startIcon={<FileDownloadIcon />}
+                  variant="contained"
+                >
+                  Export All Data
+                </Button>
+              </Box>)}
+
+          />}
+
       </div>
 
       {
