+++ frontend/src/component/Tables/PMC/PMCTable.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useState ,useMemo} from 'react';
 import { Link } from 'react-router-dom';
 import Styles from './PMCTable.module.css'
 import axios from 'axios'
@@ -7,21 +7,40 @@ import MaterialTable from 'material-table';
 import { Paper } from '@material-ui/core';
 import Modal from '../../Layout/Modal/Modal';
 import ArchitectEditForm from '../../Forms/ArchitectEditForm';
-// import MistryEditForm from '../../Forms/MistryEditForm';
 import PMCEditForm from '../../Forms/PMCEditForm'
 import { toast, ToastContainer } from 'react-toastify'
 import Select from 'react-select'
 import TextField from '@mui/material/TextField';
-
-const PMCTable = ({ modalHandler, refresh }) => {
+import FileDownloadIcon from '@mui/icons-material/FileDownload';
+import MaterialReactTable from 'material-react-table';
+import { ExportToCsv } from 'export-to-csv';
+import { dateformater } from '../Utils/util';
+import {
+  Box,
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogTitle,
+  IconButton,
+  MenuItem,
+  Stack,
+  // TextField,
+  Tooltip,
+} from '@mui/material';
+import { Delete, Edit } from '@mui/icons-material';
+const PMCTable = ({ modalHandler, refresh,isOpen }) => {
   const [PMC, setPMC] = useState([]);
   const [editModal, setEditModal] = useState(false);
   const [editModalData, setEditModalData] = useState({});
   const [tabledata, setTableData] = useState([])
+  const [originalData, setOriginalData] = useState([])
   const [startDate, setStartDate] = useState(new Date('2022-08-01'));
   const [endDate, setEndDate] = useState(new Date());
   const [branches, setBranches] = useState([]);
-  let selectedBranch = [];
+  let [selectedBranch,setSelectedBranch] = useState(null);
+  let [selectedSalesman,setSelectedSalesman] = useState(null);
+
   const [isLoading, setIsLoading] = useState(false)
   const startDateHandler = (e) => {
     setStartDate(new Date(e.target.value));
@@ -31,17 +50,11 @@ const PMCTable = ({ modalHandler, refresh }) => {
     setEndDate(new Date(e.target.value));
   }
 
-  const dateformater = (date) => {
-    let year = date.getFullYear();
-    let month = (date.getMonth() + 1) > 9 ? date.getMonth() + 1 : '0' + date.getMonth();
-    let day = (date.getDay() + 1) > 9 ? date.getDay() + 1 : '0' + date.getDay();
-    return `${year}-${month}-${day}`
-  }
 
   const submitDateRangeHandler = (e) => {
     console.log(startDate, endDate);
     let data = PMC.filter((item) => {
-      let date = item.date;
+      let date = item.date ? item.date : '01/01/1799';
       date = new Date(date);
       if (date < endDate && date > startDate) {
         return true
@@ -53,15 +66,60 @@ const PMCTable = ({ modalHandler, refresh }) => {
     setTableData(data)
   }
 
+  const columns = useMemo(
+    () => [
+      { header: 'Date', accessorKey: 'date', type: "date", dateSetting: { locale: "en-GB" },  Cell: ({cell})=>(dateformater(cell.getValue())) },
+      { header: 'Name', accessorKey: 'name' },
+      { header: 'Address', accessorKey: 'address' },
+      { header: 'Mobile Number', accessorKey: 'mobileno' },
+    ],
+    [],
+  );
+  const ops = [
+    { header: 'Date', accessorKey: 'date', type: "date", dateSetting: { locale: "en-GB" }, },
+    { header: 'Name', accessorKey: 'name' },
+    { header: 'Address', accessorKey: 'address' },
+    { header: 'Mobile Number', accessorKey: 'mobileno' },
+    // { header: 'Email', accessorKey: 'Email', },
+    // { header: 'Company_Name', accessorKey: 'companyName', },
+    // { header: 'Birth_Date', accessorKey: 'birthdate', },
+    // { header: 'Marriage_Date', accessorKey: 'marriagedate', },
+    // { header: 'Remarks', accessorKey: 'remarks', },
+    // { header: 'Bank_Name', accessorKey: 'bankname', },
+    // { header: 'IFS_Code', accessorKey: 'IFSCcode', },
+    // { header: 'Branch_Name', accessorKey: 'branchname', },
+    // { header: 'Adhar_Card', accessorKey: 'adharcard', },
+    // { header: 'Pan_Card', accessorKey: 'pancard', columnVisibility: 'false' },
+  ]
+  const csvOptions = {
+    fieldSeparator: ',',
+    quoteStrings: '"',
+    decimalSeparator: '.',
+    showLabels: true,
+    useBom: true,
+    useKeysAsHeaders: false,
+    headers: ops.map((c) => c.header),
+  };
+  const csvExporter = new ExportToCsv(csvOptions);
+  const handleExportData = () => {
 
+    csvExporter.generateCsv(tabledata);
+  };
+  const handleExportRows = (rows) => {
+    csvExporter.generateCsv(rows.map((row) => row.original));
+  };
   const delteHandler = async (id) => {
-    const data = await axios.delete(`/api/v1/pmc/delete/${id}`);
-    fetchPMC();
+    // eslint-disable-next-line no-restricted-globals
+    if(window.confirm("Are you sure ?")){
+      const data = await axios.delete(`/api/v1/pmc/delete/${id}`);
+      fetchPMC();
+    }
   }
 
   const fetchPMC = async () => {
     const { data } = await axios.get("/api/v1/pmc/getall");
     setPMC(data.pmcs);
+    setOriginalData(data.pmcs)
     setTableData(data.pmcs);
   }
 
@@ -96,13 +154,10 @@ const PMCTable = ({ modalHandler, refresh }) => {
     setIsLoading(false);
   }
   const handlebranch = (selected) => {
-    console.log(selected);
-    // setselectedBranch(selected);
-    selectedBranch = selected;
-    fetchPMCsofBranch();
+    setSelectedBranch(selected.value)
+    fetchFilteredPMC(selectedSalesman,selected.value);
   }
   const [salesman, setSalesman] = useState([]);
-  let selectedSalesman = [];
   const fetchSalesmen = async () => {
     const { data } = await axios.get("/api/v1/salesman/getall");
 
@@ -116,24 +171,45 @@ const PMCTable = ({ modalHandler, refresh }) => {
     ))
     setSalesman(salesmen);
   }
-  const fetchArchitectsofSalesman = async () => {
-    setIsLoading(true);
-    sleep(500);
+  
+  const fetchFilteredPMC =(salesman, branch) => {
 
-    // console.log(selectedSalesman);
-    const response = await axios.post("/api/v1/salesman/pmc", selectedSalesman, { headers: { "Content-Type": "application/json" } });
+    let filteredData = originalData.filter((item)=>{
+      let isBranch = false;
+      let isSalesman = false;
+      
+      item.branches.forEach((branchObject)=>{
+        if(Object.values(branchObject).includes(branch) || branch===null){
+        isBranch = true;
+      }})
+      item.salesmen.forEach((salesmanObj)=>{
+        if(Object.values(salesmanObj).includes(salesman) || salesman===null){
+          isSalesman = true;
+        }})
 
-    // console.log(response);
-    const newarchitects = response.data.pmcs;
+      console.log(isBranch, isSalesman)
+      if(isSalesman && isBranch){
+        return true
+      }
+    })
+    console.log(filteredData);
+    let data = filteredData.map((item)=>{
+      let formateddate = item.date ? item.date : '01/01/1799';
+      return {
+        date:formateddate,
+        name:item.name,
+        address:item.address,
+        mobileno:item.mobileno,
+      }
+      })
 
-    setTableData(newarchitects);
-    setIsLoading(false);
+    setPMC(data);
+    setTableData(data);  
   }
+  
   const handlesalesman = (selected) => {
-    console.log(selected);
-
-    selectedSalesman = selected;
-    fetchArchitectsofSalesman();
+    setSelectedSalesman(selected.value);
+    fetchFilteredPMC(selected.value, selectedBranch);
   }
 
   useEffect(() => {
@@ -144,7 +220,7 @@ const PMCTable = ({ modalHandler, refresh }) => {
   const handleCallbackCreate = (childData) => {
     // console.log("Parent Invoked!!")
     toast.success("PMC edited");
-    fetchPMC();
+    // fetchPMC();
   }
   const customStyles = {
     control: base => ({
@@ -223,75 +299,98 @@ const PMCTable = ({ modalHandler, refresh }) => {
           </div>
         </div>
 
-        {PMC && <MaterialTable
-          isLoading={isLoading}
-          className={Styles.Table}
-          columns={[
-            { title: 'Date', field: 'date', type: "date", dateSetting: { locale: "en-GB" }, },
-            { title: 'Name', field: 'name' },
-            { title: 'Email', field: 'Email', hidden: 'true' },
-            { title: 'Address', field: 'address' },
-            { title: 'Company Name', field: 'companyName', hidden: 'true' },
-            { title: 'Birth Date', field: 'birthdate', hidden: 'true' },
-            { title: 'Marriage Date', field: 'marriagedate', hidden: 'true' },
-            { title: 'Remarks', field: 'remarks', hidden: 'true' },
-            { title: 'Bank Name', field: 'bankname', hidden: 'true' },
-            { title: 'IFS Code', field: 'IFSCcode', hidden: 'true' },
-            { title: 'Branch Name', field: 'branchname', hidden: 'true' },
-            { title: 'Adhar Card', field: 'adharcard', hidden: 'true' },
-            { title: 'Pan Card', field: 'pancard', hidden: 'true' },
-          ]}
-          data={tabledata}
-          options={{
-            sorting: true,
-            headerStyle: {
-              zIndex: 0
-            },
-            showTitle: false,
-            actionsColumnIndex: -1,
-            filtering: true,
-            exportButton:true
-          }}
-          components={{
-            Container: props => <Paper {...props}
-              elevation={0}
-              style={{
-                padding: 20,
-                width: "100%",
-              }} />
-          }}
+        {PMC &&
+          <MaterialReactTable
+            displayColumnDefOptions={{
+              'mrt-row-actions': {
+                muiTableHeadCellProps: {
+                  align: 'center',
+                },
 
-          actions={[
-            {
-              icon: 'edit',
-              tooltip: 'Edit',
-              onClick: (event, rowData) => {
-                window.scrollTo({
-                  top: 0,
-                  left: 0,
-                  behavior: "smooth"
-                });
-                setEditModalData(rowData);
-                setEditModal(true);
-                console.log(`Edit `, rowData)
-              }
-            },
-            {
-              icon: 'delete',
-              tooltip: 'Delete',
-              onClick: (event, rowData) => {
-                window.scrollTo({
-                  top: 0,
-                  left: 0,
-                  behavior: "smooth"
-                });
-                // Do save operation
-                delteHandler(rowData._id);
-                console.log(`delete `, rowData)
-              }
+                size: 120,
+              },
+            }}
+
+            muiTopToolbarProps={
+              ({ }) => ({
+                color: 'green',
+                sx: { display: 'block' },
+                zIndex: '0'
+              })
             }
-          ]}
-        />}
+            columns={columns}
+            data={tabledata}
+            enableEditing
+            enableRowNumbers
+            rowNumberMode='original'
+            enableTopToolbar={!editModal && !isOpen}
+
+            muiTablePaginationProps={{
+              rowsPerPageOptions: [5, 10],
+              showFirstLastPageButtons: true,
+            }}
+            enableGlobalFilter={true}
+            positionActionsColumn='last'
+            renderRowActions={({ row, table }) => (
+              <Box sx={{ display: 'flex', gap: '1rem' }}>
+                <Tooltip arrow placement="left" title="Edit">
+                  <IconButton onClick={() => {
+                    window.scrollTo({
+                      top: 0,
+                      left: 0,
+                      behavior: "smooth"
+                    });
+
+                    setEditModalData(row.original)
+                    setEditModal(true);
+                  }}>
+                    <Edit />
+
+                  </IconButton>
+                </Tooltip>
+                <Tooltip arrow placement="right" title="Delete">
+                  <IconButton color="error" onClick={() => {
+                    window.scrollTo({
+                      top: 0,
+                      left: 0,
+                      behavior: "smooth"
+                    });
+                    delteHandler(row.original._id);
+                    console.log(`delete `, row)
+                  }}>
+                    <Delete />
+                  </IconButton>
+                </Tooltip>
+              </Box>
+            )}
+            renderTopToolbarCustomActions={({ table }) => (
+              <Box
+                sx={{ display: 'flex', gap: '1rem', p: '0.5rem', flexWrap: 'wrap' }}
+              >
+                <Button
+                  disabled={table.getPrePaginationRowModel().rows.length === 0}
+
+                  onClick={() =>
+                    handleExportRows(table.getPrePaginationRowModel().rows)
+                  }
+                  startIcon={<FileDownloadIcon />}
+                  variant="contained"
+                >Export All Rows</Button>
+                <Button
+                  className={Styles.bu}
+                  color="primary"
+                  onClick={handleExportData}
+                  startIcon={<FileDownloadIcon />}
+                  variant="contained"
+                >
+                  Export All Data
+                </Button>
+              </Box>)}
+
+          />}
+
+
+
 
       </div>
